% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_immune.R
\name{immune_initial}
\alias{immune_initial}
\title{Within-host Pre-erythrocytic Immune ODE Model}
\usage{
immune_initial(a, zeta, psi, theta, ...)
}
\arguments{
\item{a}{age of individual}

\item{zeta}{individual biting heterogeneity}

\item{psi}{relative biting rate (used to calculate psi(a))}

\item{theta}{vector of model parameters (see \code{\link{RACD_Parameters}})}

\item{...}{additional parameters passed to numerical integrator \code{\link[deSolve]{ode}}}
}
\description{
Calculate an individual's pre-erythrocytic immunity at age a for given EIR heterogeneities (biting rate & geographic location).
This function calls \code{\link{immune_ode}}
}
