% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_infection.R
\name{infection_ode}
\alias{infection_ode}
\title{Within-host Infection ODE Model}
\usage{
infection_ode(time, theta, state, ...)
}
\arguments{
\item{time}{times to evaluate state variables at}

\item{theta}{named vector of parameters}

\item{state}{initial states}

\item{...}{additional parameters passed to numerical integrator \code{\link[deSolve]{ode}}}
}
\description{
write me
}
