% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACD-Setup.R
\name{RACD_Setup}
\alias{RACD_Setup}
\title{Initialize State for RACD Model}
\usage{
RACD_Setup(theta)
}
\arguments{
\item{theta}{named vector of parameters (see \code{\link{RACD_Parameters}})}
}
\description{
do something
}
