% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACD-Output.R
\name{RACD_StateVector}
\alias{RACD_StateVector}
\title{Process RACD Simulation Output}
\usage{
RACD_StateVector(outfile)
}
\arguments{
\item{outfile}{path to logged events}
}
\description{
Given output from \code{\link[RACD]{RACD_Simulation}}, take continuous time state transition output
and make state occupancy vector by binning events by day.
}
